Add files via upload
It covers essential plotting techniques, including line and scatter plots, along with advanced customization options like titles, labels, and legends. The notebook also demonstrates how to work with multiple figures and subplots and integrates Matplotlib with Pandas and NumPy. Additionally, it includes examples of styling plots and exporting them in various formats. This resource is ideal for anyone looking to master data visualization with Matplotlib.